+++ docs/conf.py
@@ -281,7 +281,7 @@ intersphinx_mapping = {
     "flask": ("http://flask.pocoo.org/docs/", None),
     "flask_login": ("https://flask-login.readthedocs.io/en/latest/", None),
     "werkzeug": ("http://werkzeug.pocoo.org/docs/", None),
-    "requests": ("https://requests.kennethreitz.org/en/latest/", None),
+    "requests": ("http://docs.python-requests.org/en/latest/", None),
     "oauthlib": ("https://oauthlib.readthedocs.io/en/latest/", None),
     "requests_oauthlib": ("https://requests-oauthlib.readthedocs.io/en/latest/", None),
     "sqlalchemy": ("https://docs.sqlalchemy.org/en/latest/", None),