+++ CHANGELOG.rst
@@ -3,7 +3,8 @@ Changelog
 
 `unreleased`_
 -------------
-Nothing yet
+* Updated Meetup pre-set configuration to include the ``client_id`` in the
+  OAuth token request.
 
 `2.2.0`_ (2019-06-04)
 ---------------------