+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
   to include the ``client_id`` in the OAuth token request.
 * Removed Okta pre-set configuration, since it doesn't add any value over
   using ``OAuth2ConsumerBlueprint`` directly.
+* Updated Azure to allow defining ``authorization_url_params`` 
 
 `2.2.0`_ (2019-06-04)
 ---------------------